add ts types for pagenavigation.js
